requirements.txt
gunicorn
requests
flask-socketio
dill
numpy
scipy
scikit-learn
flask
google-cloud-speech
eventlet